refactor(QnAService) : 리팩토링 할 더미서비스객체 생성
